Solve the conflict
